+++ src/pages/Folder/components/FoldersContainer/FoldersContainer.types.ts
@@ -1,0 +1,9 @@
+export interface Folder {
+  id: number;
+  name: string;
+  created_at?: string;
+  user_id?: number;
+  link?: {
+    count: number;
+  };
+}
